requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
gspread
pandas
google-auth-oauthlib
huggingface-hub
pydantic-settings
numpy
pinecone>=3.0.0  # <-- UPDATE THIS LINE
psycopg2-binary